+++ malachite-float/src/comparison/partial_cmp_abs_integer.rs
@@ -57,7 +57,7 @@ impl PartialOrdAbs<Integer> for Float {
             } else if *e_x <= 0 {
                 Less
             } else {
-                e_x.unsigned_abs()
+                u64::from(e_x.unsigned_abs())
                     .cmp(&other.significant_bits())
                     .then_with(|| x.cmp_normalized(y.unsigned_abs_ref()))
             }),
