+++ malachite-float/src/comparison/partial_cmp_integer.rs
@@ -63,7 +63,7 @@ impl PartialOrd<Integer> for Float {
                 } else if *e_x <= 0 {
                     Less
                 } else {
-                    e_x.unsigned_abs()
+                    u64::from(e_x.unsigned_abs())
                         .cmp(&other.significant_bits())
                         .then_with(|| x.cmp_normalized(y.unsigned_abs_ref()))
                 };
