+++ malachite-float/src/comparison/partial_eq_integer.rs
@@ -48,7 +48,7 @@ impl PartialEq<Integer> for Float {
                 *other != 0u32
                     && *sign == (*other >= 0u32)
                     && *exponent >= 0
-                    && other.significant_bits() == exponent.unsigned_abs()
+                    && other.significant_bits() == u64::from(exponent.unsigned_abs())
                     && significand.cmp_normalized(other.unsigned_abs_ref()) == Equal
             }
             _ => false,
